add vercel

+++ build_files.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-#!/bin/bash
 echo "Where is Python:"
 which python
 echo "Python Version:"
@@ -9,5 +8,16 @@ which pip
 echo "Pip Version:"
 pip --version
 
-pip install -r requirements.txt
-python manage.py collectstatic --no-input
+# Проверка на наличие файла requirements.txt перед установкой зависимостей
+if [ -f "requirements.txt" ]; then
+    pip install -r requirements.txt
+else
+    echo "requirements.txt not found"
+fi
+
+# Проверка на наличие manage.py перед выполнением Django команд
+if [ -f "manage.py" ]; then
+    python manage.py collectstatic --no-input
+else
+    echo "manage.py not found"
+fi
